[ci] Initial framework refactor
This commit ensures we use framework
for testing both wmcb and wsu. In future,
when we move to operator based model,
we can run both the tests in a single
shot.
