+++ hack/run-wmcb-ci-e2e-test.sh
@@ -5,14 +5,13 @@ set -o pipefail
 
 # TODO: Add steps to execute code that will kick off the e2e tests in CI
 WMCO_ROOT=$(dirname "${BASH_SOURCE}")/..
-INTERNAL_TEST_DIR=$WMCO_ROOT/internal/test/
+WMCB_TEST_DIR=$WMCO_ROOT/internal/test/wmcb
 
 # Build the unit test binary
 cd "${WMCO_ROOT}"
 make build-wmcb-unit-test
 
 # Transfer the binary and run the unit tests
-cd "${INTERNAL_TEST_DIR}"
-CGO_ENABLED=0 GO111MODULE=on go test -c wmcb_test.go -o wmcb_framework
-./wmcb_framework -test.run=TestWMCBUnit -binaryToBeTransferred=../../wmcb_unit_test.exe
+cd "${WMCB_TEST_DIR}"
+CGO_ENABLED=0 GO111MODULE=on go test -v -run=TestWMCBUnit -binaryToBeTransferred=../../../wmcb_unit_test.exe -timeout=30m .
